lesson 33 cookie

+++ src/app/api/users/profile/[id]/route.ts
@@ -2,6 +2,7 @@ import { NextRequest, NextResponse } from 'next/server';
 import prisma from '@/utils/db';
 import jwt from 'jsonwebtoken';
 import { JWTPayload } from '@/utils/types';
+import { verifyToken } from '@/utils/verifyToken';
 interface Props {
     params: { id: string }
 }
@@ -20,11 +21,10 @@ export async function DELETE(request: NextRequest, { params }: Props) {
             return NextResponse.json({ message: 'User Not Found' }, { status: 404 });
         }
 
-        const authToken:string = request.headers.get('authToken') as string;
-        const userFromToken = jwt.verify(authToken, process.env.JWT_SECRET as string) as JWTPayload;
-        if (userFromToken.id === user.id) {
+        const userFromToken = verifyToken(request);
+        if (userFromToken !==null && userFromToken.id === user.id) {
             await prisma.user.delete({ where: { id: parseInt(params.id) } });
-            return NextResponse.json({ message: 'User deleted successfully' }, { status: 200 });
+            return NextResponse.json({ message: 'Your Profile ( account ) has been deleted ' }, { status: 200 });
         }
 
         return NextResponse.json({ message: 'Only User Himself Can Delete His Profile , Forbidden ' }, { status: 403 }); // 403 Forbidden
